% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-img.R
\name{arrumar}
\alias{arrumar}
\title{Arruma o arquivo num BD pronto para predição.}
\usage{
arrumar(arqs_treino, nm = character(0), spread = TRUE)
}
\arguments{
\item{arqs_treino}{arquivo com a imagem em jpeg.}

\item{nm}{nomes do modelo para completar a base de predição.}

\item{spread}{dar spread?}
}
\description{
Deve ser usada dentro da função predizer.
}

