% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{decifrar}
\alias{decifrar}
\title{Decifra o audio}
\usage{
decifrar(arq, deletar = TRUE)
}
\arguments{
\item{arq}{arquivo que contém o audio.}

\item{deletar}{se TRUE, deleta o arquivo .wav gerado. Se FALSE, deixa
o arquivo na pasta getwd().}
}
\value{
character vector de tamanho 1 com as letras preditas.
}
\description{
Le um arquivo de audio e tenta decifrar quais sao as letras.
}

