% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{ler}
\alias{ler}
\title{Le e processa um arquivo mp3}
\usage{
ler(a, salvar = FALSE, path = "data-raw", deletar = TRUE)
}
\arguments{
\item{a}{arquivo mp3 a ser lido}

\item{salvar}{se TRUE, salva arquivos .wav e .rds com os trechos do audio.
Se FALSE, não faz nada.}

\item{path}{pasta para salvar os arquivos caso salvar == TRUE.}

\item{deletar}{se TRUE, deleta o arquivo .wav gerado. Se FALSE, deixa
o arquivo na pasta getwd().}
}
\value{
um vetor de inteiros com a soma dos valores absolutos das frequencias
de cada trecho identificado. Esses numeros podem ser considerados como
uma fingerprint da letra, pois nenhuma outra letra possui a mesma soma
de frequencias.
}
\description{
Le um arquivo mp3, converte para .wav com mpg123 (instale!), depois
usa o pacote \code{tuneR} para ler o arquivo. Pode ter vários side effects.
}

