% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-new.R
\name{ocr}
\alias{ocr}
\title{Instalar ocR}
\usage{
ocr(dados)
}
\description{
https://github.com/greenore/ocR
}
\details{
criar o arquivo
/usr/share/tesseract-ocr/tessdata/configs
e adicionar o seguinte conteudo
tessedit_char_whitelist abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ
}

